+++ typescript/scheme/scheme.ts
@@ -115,8 +115,16 @@ export class Scheme {
     return this.props.outline;
   }
 
+  get outlineVariant(): number {
+    return this.props.outlineVariant;
+  }
+
   get shadow(): number {
     return this.props.shadow;
+  }
+
+  get scrim(): number {
+    return this.props.scrim;
   }
 
   get inverseSurface(): number {
@@ -191,7 +199,9 @@ export class Scheme {
       surfaceVariant: core.n2.tone(90),
       onSurfaceVariant: core.n2.tone(30),
       outline: core.n2.tone(50),
+      outlineVariant: core.n2.tone(80),
       shadow: core.n1.tone(0),
+      scrim: core.n1.tone(0),
       inverseSurface: core.n1.tone(20),
       inverseOnSurface: core.n1.tone(95),
       inversePrimary: core.a1.tone(80)
@@ -226,7 +236,9 @@ export class Scheme {
       surfaceVariant: core.n2.tone(30),
       onSurfaceVariant: core.n2.tone(80),
       outline: core.n2.tone(60),
+      outlineVariant: core.n2.tone(30),
       shadow: core.n1.tone(0),
+      scrim: core.n1.tone(0),
       inverseSurface: core.n1.tone(90),
       inverseOnSurface: core.n1.tone(20),
       inversePrimary: core.a1.tone(40)
@@ -257,7 +269,9 @@ export class Scheme {
     surfaceVariant: number,
     onSurfaceVariant: number,
     outline: number,
+    outlineVariant: number,
     shadow: number,
+    scrim: number,
     inverseSurface: number,
     inverseOnSurface: number,
     inversePrimary: number